Renaming to be able to create a new pypi package
