+++ Makefile
@@ -18,7 +18,7 @@ clean-pyc:
 .PHONY: pep8
 # target: pep8 - Check code for pep8 rules
 pep8:
-	@flake8 mattermost_bot --exclude=mattermost_bot/settings.py
+	@flake8 mmpy_bot --exclude=mmpy_bot/settings.py
 
 .PHONY: release
 # target: release - Release app into PyPi
@@ -33,7 +33,7 @@ sdist: clean
 .PHONY: run
 # target: run - Run bot
 run:
-	matterbot
+	mmpy_bot
 
 .PHONY: sphinx
 # target: sphinx - Make app docs
