+++ docs/conf.py
@@ -2,7 +2,7 @@
 
 import os
 
-from mattermost_bot import get_version
+from mmpy_bot import get_version
 
 os.environ['DJANGO_SETTINGS_MODULE'] = 'settings'
 
@@ -10,24 +10,24 @@ extensions = ['sphinx.ext.autodoc', 'sphinx.ext.todo']
 templates_path = ['_templates']
 source_suffix = '.rst'
 master_doc = 'index'
-project = u'mattermost_bot'
+project = u'mmpy_bot'
 copyright = u'2016, '
 version = get_version()
 release = get_version()
 exclude_patterns = []
 pygments_style = 'sphinx'
 html_theme = 'default'
-htmlhelp_basename = 'mattermost_botdoc'
+htmlhelp_basename = 'mmpy_botdoc'
 latex_documents = [
-    ('index', 'mattermost_bot.tex', u'mattermost_bot Documentation',
+    ('index', 'mmpy_bot.tex', u'mmpy_bot Documentation',
      u'', 'manual'),
 ]
 man_pages = [
-    ('index', 'mattermost_bot', u'mattermost_bot Documentation',
+    ('index', 'mmpy_bot', u'mmpy_bot Documentation',
      [u'gotlium'], 1)
 ]
 texinfo_documents = [
-    ('index', 'mattermost_bot', u'Mattermost-bot Documentation',
-     u'gotlium', 'mattermost_bot', 'One line description of project.',
+    ('index', 'mmpy_bot', u'Mattermost-bot Documentation',
+     u'gotlium', 'mmpy_bot', 'One line description of project.',
      'Miscellaneous'),
 ]
