+++ docs/requirements.txt
@@ -1,2 +1,2 @@
 Django>1.8,<1.9
-mattermost_bot>1.0.9
+mmpy_bot>1.0.9
