+++ mmpy_bot/__init__.py
@@ -1,4 +1,4 @@
-VERSION = (1, 0, 20)
+VERSION = (1, 0, 0)
 
 
 def get_version():
