+++ mmpy_bot/plugins/access.py
@@ -1,8 +1,8 @@
 # -*- encoding: utf-8 -*-
 
-from mattermost_bot.utils import allow_only_direct_message
-from mattermost_bot.utils import allowed_users
-from mattermost_bot.bot import respond_to
+from mmpy_bot.utils import allow_only_direct_message
+from mmpy_bot.utils import allowed_users
+from mmpy_bot.bot import respond_to
 
 
 @respond_to('^admin$')
