+++ mmpy_bot/plugins/sleep.py
@@ -2,7 +2,7 @@
 
 import time
 
-from mattermost_bot.bot import respond_to
+from mmpy_bot.bot import respond_to
 
 
 @respond_to('sleep (.*)')
