+++ tests/unit_tests/local_plugins/hello.py
@@ -2,8 +2,8 @@
 
 import re
 
-from mattermost_bot.bot import listen_to
-from mattermost_bot.bot import respond_to
+from mmpy_bot.bot import listen_to
+from mmpy_bot.bot import respond_to
 
 
 @respond_to('hello$', re.IGNORECASE)
