+++ tests/unit_tests/single_plugin/mock_plugin.py
@@ -1,6 +1,6 @@
-from mattermost_bot.utils import allow_only_direct_message
-from mattermost_bot.utils import allowed_users
-from mattermost_bot.bot import respond_to
+from mmpy_bot.utils import allow_only_direct_message
+from mmpy_bot.utils import allowed_users
+from mmpy_bot.bot import respond_to
 
 @respond_to('^admin$')
 @allow_only_direct_message()
